ref(cli): move RiversConfig to separate module
